# docs/msc2/api-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MSC 2 v1 API contract",
    "version": "v1",
    "description": "The versioned HTTP contract MSC 2's skeletal agent (Phase 2) and every later phase implement against. Assembled per docs/msc2/rolling-plan.md P2.8 from the P0.23/P0.30/P0.32 MSC 1 baseline (docs/msc2/api-baseline/openapi.json) plus this phase's own designs: the /v1/ namespace and ErrorDTO envelope (P2.4, versioning-and-errors.md), the operation routes and OperationDTO (P2.5, operation-model.md), the capabilities route (P2.6, capability-model.md), helpId placement (P2.2, helpid-contract.md), and the permission-category annotation on every route (P2.1, permission-vocabulary.csv). The console and operation-progress WebSocket channels (P2.7) are documented separately in docs/msc2/api-contract/websocket-v1.json, matching the baseline's own precedent (P0.24) of keeping WS channels out of the OpenAPI paths object. Everything sourced from P2.4/P2.5/P2.6/P2.2's own documents carries their Proposed/Confirmed status forward -- this file does not itself promote anything to Approved."
  },
  "paths": {
...
    "/v1/operations": {
      "post": {
        "summary": "Create a long-running operation",
        "x-permission-category": "serverControl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationCreateRequestDTO"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted for asynchronous processing. Always state: queued at creation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationDTO"
                }
              }
            }
          },
          "400": {
            "description": "Unrecognized type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "invalid_body"
          }
        },
        "x-notes": "P2.5 SS4.1. This phase, the only type the skeletal agent accepts is demo-install."
      }
    },
    "/v1/operations/{id}": {
      "get": {
        "summary": "Read an operation's current state",
        "x-permission-category": "none",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current OperationDTO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationDTO"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id -- this phase, includes any operation forgotten across a restart (no journal yet).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "not_found",
            "x-help-id": "operations.not-found"
          }
        },
        "x-notes": "P2.5 SS4.2."
      }
    },
    "/v1/operations/{id}/cancel": {
      "post": {
        "summary": "Request cancellation of an operation",
        "x-permission-category": "serverControl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated OperationDTO, state: cancelled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationDTO"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "not_found"
          },
          "409": {
            "description": "Operation is already terminal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDTO"
                }
              }
            },
            "x-error-code": "conflict",
            "x-help-id": "operations.cancel-not-legal"
          }
        },
        "x-notes": "P2.5 SS4.3."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDTO": {
        "type": "object",
        "description": "One error envelope for every non-2xx response across every /v1/ route (P2.4 SS5-6, a D-006-point-3 correction unifying the baseline's split Error/typed-DTO failure pattern).",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable snake_case identifier, e.g. not_found, conflict, invalid_body."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, iOS-visible text."
          },
          "helpId": {
            "type": "string",
            "nullable": true,
            "description": "Optional pointer into GET /v1/help/{helpId} (P2.2)."
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Optional free-form structured context."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "OperationDTO": {
        "type": "object",
        "description": "P2.5 SS2 -- the wire shape for a long-running operation.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Opaque, server-generated. Never client-supplied."
          },
          "type": {
            "type": "string",
            "description": "Kind of work, e.g. demo-install. Not a closed enum -- new values are additive (P2.5 SS2)."
          },
          "target": {
            "type": "string",
            "nullable": true,
            "description": "The thing the operation acts on, typically a server name/ID. Null if the operation type has no natural target."
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "description": "P2.5 SS3's closed state machine."
          },
          "progress": {
            "type": "object",
            "nullable": true,
            "properties": {
              "current": {
                "type": "integer",
                "minimum": 0
              },
              "total": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "current",
              "total"
            ],
            "description": "Null while queued, or for a type with no natural countable unit."
          },
          "statusLine": {
            "type": "string",
            "nullable": true,
            "description": "Human-readable, meant for direct display."
          },
          "result": {
            "type": "object",
            "nullable": true,
            "description": "Present only when state == succeeded. Shape defined per type."
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDTO",
            "nullable": true,
            "description": "Present only when state == failed."
          }
        },
        "required": [
          "id",
          "type",
          "state"
        ]
      },
      "OperationCreateRequestDTO": {
        "type": "object",
        "description": "P2.5 SS4.1 -- POST /v1/operations request body.",
        "properties": {
          "type": {
            "type": "string",
            "description": "Required. Unrecognized values are a 400 invalid_body."
          },
          "target": {
            "type": "string",
            "nullable": true
          },
          "params": {
            "type": "object",
            "nullable": true,
            "description": "Free-form, shape defined per type."
          }
        },
        "required": [
          "type"
        ]
      },
...
    }
  }
}